feature: Calculate and show BMI on weights page #17

+++ Fytonyashka/DataModels/UserProfileInputModel.cs
@@ -15,7 +15,7 @@ public class UserProfileInputModel
     [Range(0.01, double.MaxValue, ErrorMessage = "Height must be greater than 0")]
     public int Height { get; set; } = 0;
 
-    public int? Weight { get; set; } = 0;
+    public double? Weight { get; set; } = 0;
 
     [Required(ErrorMessage = "Username is required")]
     [StringLength(20, MinimumLength = 3, ErrorMessage = "Minimum length of Username is 3 and maximum is 20")]
